clean up to make it work
