+++ package.json
@@ -42,6 +42,7 @@
     "gettext-parser": "1.2.2",
     "github": "12.0.1",
     "glob": "7.1.2",
+    "googleapis": "23.0.0",
     "jest": "20.0.4",
     "lint-staged": "4.0.1",
     "mkdirp": "0.5.1",
