+++ src/commands/deployUtils/deploy.js
@@ -1,26 +1,5 @@
-const chalk = require('chalk');
-const cloudApi = require('./cloudApi');
-
 module.exports = (config, { versions, operations }) => {
-	const { auth, appsId, servicesId, version, versionIds } = config;
-	const logOp = id => op => {
-		console.log(
-			chalk.green(`Creating version ${id} for ${op.metadata.user}\n`),
-			chalk.yellow(`Operation ${op.name} in progress`)
-		);
-		return op;
-	};
-	const createVersion = id =>
-		cloudApi.versions
-			.create({
-				auth,
-				appsId,
-				servicesId,
-				resource: versions.spec(id),
-			})
-			.then(logOp(id))
-			.then(operations.version);
-
+	const { version, versionIds } = config;
 	return {
 		create: () => {
 			const { noExisting, noNewer } = versions.validate;
@@ -29,7 +8,7 @@ module.exports = (config, { versions, operations }) => {
 				`versions ${versionIds.join(', ')}`
 			);
 			return Promise.all([noExisting(), noNewer()]).then(() =>
-				Promise.all(versionIds.map(createVersion))
+				Promise.all(versionIds.map(versions.create))
 			);
 		},
 		del: () =>
